third_party/webgl/src/conformance-suites/1.0.1/conformance/misc/bad-arguments-test.html

<!--
Copyright (C) 2011 Apple Computer, Inc.  All rights reserved.

Redistribution and use in source and binary forms, with or without
modification, are permitted provided that the following conditions
are met:
1. Redistributions of source code must retain the above copyright
   notice, this list of conditions and the following disclaimer.
2. Redistributions in binary form must reproduce the above copyright
   notice, this list of conditions and the following disclaimer in the
   documentation and/or other materials provided with the distribution.

THIS SOFTWARE IS PROVIDED BY APPLE COMPUTER, INC. ``AS IS'' AND ANY
EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE COMPUTER, INC. OR
CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->
<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
var wtu = WebGLTestUtils;
description("Tests calling WebGL APIs with wrong argument types");

var context = wtu.create3DContext();
var program = wtu.loadStandardProgram(context);
var shader = wtu.loadStandardVertexShader(context);

function mayThrow(expression) {
  wtu.shouldThrowOrGenerateGLError(context, [context.NO_ERROR, context.INVALID_VALUE, context.INVALID_OPERATION, context.INVALID_ENUM], expression);
}

assertMsg(program != null, "Program Compiled");
assertMsg(shader != null, "Shader Compiled");

var loc = context.getUniformLocation(program, "u_modelViewProjMatrix");
assertMsg(loc != null, "getUniformLocation succeeded");

var arguments = [
  { value: "foo",
    throws: "yes" },
  { value: 0,
    throws: "yes" },
  { value: null,
    throws: "maybe" },
  { value: undefined,
    throws: "maybe" }
];

var argument;

for (var i = 0; i < arguments.length; ++i) {
  var func, func2;
  if (arguments[i].throws == "yes") {
    func = shouldThrow;
    func2 = shouldThrow;
  } else {
    func = mayThrow;
    func2 = shouldBeUndefined;
  }
  argument = arguments[i].value;
  func("context.compileShader(argument)");
  func("context.linkProgram(argument)");
  func("context.attachShader(program, argument)");
  func("context.attachShader(argument, shader)");
  func("context.detachShader(program, argument)");
  func("context.detachShader(argument, shader)");
  func("context.shaderSource(argument, 'foo')");
  func("context.bindAttribLocation(argument, 0, 'foo')");
  func("context.getProgramInfoLog(argument)");
  func("context.getProgramParameter(argument, 0)");
  func("context.getShaderInfoLog(argument)");
  func("context.getShaderParameter(argument, 0)");
  func("context.getShaderSource(argument)");
  func("context.getUniform(argument, loc)");
  func("context.getUniform(program, argument)");
  func("context.getUniformLocation(argument, 'u_modelViewProjMatrix')");

  func2("context.useProgram(argument)");
  func2("context.bindBuffer(context.ARRAY_BUFFER, argument)");
  func2("context.bindFramebuffer(context.FRAMEBUFFER, argument)");
  func2("context.bindRenderbuffer(context.RENDERBUFFER, argument)");
  func2("context.bindTexture(context.TEXTURE_2D, argument)");
  func2("context.framebufferRenderbuffer(context.FRAMEBUFFER, context.DEPTH_ATTACHMENT, context.RENDERBUFFER, argument)");
  func2("context.framebufferTexture2D(context.FRAMEBUFFER, context.COLOR_ATTACHMENT0, context.TEXTURE_2D, argument, 0)");
  func2("context.uniform2fv(argument, new Float32Array([0.0, 0.0]))");
  func2("context.uniform2iv(argument, new Int32Array([0, 0]))");
  func2("context.uniformMatrix2fv(argument, false, new Float32Array([0.0, 0.0, 0.0, 0.0]))");
}

successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
